+++ cargo-concordium/src/build.rs
@@ -80,7 +80,7 @@ pub fn build_contract(
         }
         utils::WasmVersion::V1 => {
             if build_schema.build() {
-                let schema = build_contract_schema(cargo_args, utils::generate_contract_schema_v1)
+                let schema = build_contract_schema(cargo_args, utils::generate_contract_schema_v2)
                     .context("Could not build module schema.")?;
                 if build_schema.embed() {
                     schema_bytes = to_bytes(&schema);